pyproject.toml
[project]
name = "ane-gpu"
version = "0.1.0"
description = "Heterogeneous ANE+GPU speculative decoding for Apple Silicon — use the Neural Engine as a draft model while the GPU runs the target LLM"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<3.14"
authors = [
    { name = "Ken Fink" },
]
keywords = ["apple-silicon", "neural-engine", "speculative-decoding", "mlx", "coreml", "llm", "inference"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mlx>=0.22",
    "mlx-lm>=0.21",
    "coremltools>=8.0",
    "transformers>=4.46",
    "torch>=2.4",
    "numpy>=1.26",
    "huggingface-hub>=0.20",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.2",
    "ruff>=0.4",
]

[project.scripts]
ane-gpu = "src.cli:main"

[project.urls]
Homepage = "https://github.com/kenfink/ane-gpu"
Repository = "https://github.com/kenfink/ane-gpu"
Issues = "https://github.com/kenfink/ane-gpu/issues"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests that load real models (deselect with '-m \"not slow\"')",
    "requires_ane: marks tests that need ANE hardware",
]
